+++ core/src/main/scala/pureconfig/DerivedReaders.scala
@@ -20,9 +20,26 @@ import shapeless.labelled._
 trait AllowMissingKey { self: ConfigReader[_] => }
 
 /**
+ * Trait extending [[DerivedReaders1]] that contains `ConfigReader` instances for `AnyVal`.
+ *
+ * This trait exists to give priority to the `AnyVal` derivation over the generic product derivation.
+ */
+trait DerivedReaders extends DerivedReaders1 {
+  implicit def deriveAnyVal[T, U](
+    implicit
+    ev: T <:< AnyVal,
+    unwrapped: Unwrapped.Aux[T, U],
+    reader: ConfigReader[U]): ConfigReader[T] =
+    new ConfigReader[T] {
+      def from(value: ConfigValue): Either[ConfigReaderFailures, T] =
+        reader.from(value).right.map(unwrapped.wrap)
+    }
+}
+
+/**
  * Trait containing `ConfigReader` instances for collection, product and coproduct types.
  */
-trait DerivedReaders {
+trait DerivedReaders1 {
 
   private[pureconfig] trait WrappedConfigReader[Wrapped, SubRepr] extends ConfigReader[SubRepr]
 
@@ -198,6 +215,7 @@ trait DerivedReaders {
       cc.value.from(config).right.map(gen.from)
     }
   }
+
 }
 
 object DerivedReaders extends DerivedReaders