+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - New features
   - `ConfigReader`, `ConfigWriter` and `ConfigConvert` now have combinators such as `map`, `flatMap` and `contramap`,
     making them easier to compose.
+- Bug fixes
+  - `Duration.Undefined` is correctly handled when reading an writing configurations. [[#184](https://github.com/melrief/pureconfig/issues/184)].
 
 ### 0.7.0 (Apr 2, 2017)