+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - New features
   - `ConfigReader`, `ConfigWriter` and `ConfigConvert` now have combinators such as `map`, `flatMap` and `contramap`,
     making them easier to compose.
+  - New mechanism to read and write [value classes](http://docs.scala-lang.org/overviews/core/value-classes.html).
+    The readers and writers of the inner type are used instead of the ones for products.
 - Bug fixes
   - `Duration.Undefined` is correctly handled when reading and writing configurations. [[#184](https://github.com/melrief/pureconfig/issues/184)]